consolidate topic breadcrumb logic
