+++ kolibri/plugins/coach/assets/src/views/reports/breadcrumbs.vue
@@ -1,7 +1,7 @@
 <template>
 
   <div class="breadcrumb-wrapper">
-    <span v-for="ancestor in list" class="crumb">
+    <span v-for="ancestor in contentBreadcrumbs" class="crumb">
       <span v-if="ancestor.vlink">
         <router-link :to="ancestor.vlink">{{ ancestor.title }}</router-link>
       </span>
@@ -14,11 +14,55 @@
 
 <script>
 
+  const CoachConstants = require('../../constants');
+
   module.exports = {
-    props: {
-      list: {
-        type: Array,
-        required: true,
+    computed: {
+      contentBreadcrumbs() {
+        return [
+          // link to the root channels page
+          {
+            title: 'Channels',
+            vlink: {
+              name: CoachConstants.PageNames.TOPIC_CHANNELS,
+              params: {
+                classId: this.pageState.classId,
+              },
+            },
+          },
+          // links to each ancestor
+          ...this.pageState.contentScopeSummary.ancestors.map((item, index) => {
+            const breadcrumb = { title: item.title };
+            if (index) {
+              // links to parent topics
+              breadcrumb.vlink = {
+                name: CoachConstants.PageNames.TOPIC_ITEM_LIST,
+                params: {
+                  classId: this.pageState.classId,
+                  channelId: this.pageState.channelId,
+                  topicId: item.id,
+                },
+              };
+            } else {
+              // link to channel root
+              breadcrumb.vlink = {
+                name: CoachConstants.PageNames.TOPIC_CHANNEL_ROOT,
+                params: {
+                  classId: this.pageState.classId,
+                  channelId: this.pageState.channelId,
+                },
+              };
+            }
+            return breadcrumb;
+          }),
+          // current item
+          { title: this.pageState.contentScopeSummary.title }
+        ];
+      },
+    },
+    vuex: {
+      getters: {
+        pageState: state => state.pageState,
       },
     },
   };
